requirements.txt
playwright==1.41.2
pandas==2.2.0
fake-useragent==1.4.0
python-dotenv==1.0.1
tenacity==8.2.3
